imdb/requirement.txt
asgiref==3.2.10
click==7.1.2
dj-database-url==0.5.0
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
Werkzeug==1.0.1
whitenoise==5.2.0
